requirements.txt
dotenv
django-cors-headers-4.7.0
djangorestframework-3.15.2


node v22.14.0
npm 10.9.2
npx 10.9.2
nvm 0.39.5
create-react-app@5.1.0


Truffle v5.11.5 (core: 5.11.5)
Ganache v7.9.1
Solidity v0.5.16 (solc-js)
Web3.js v1.10.0